hooking fixes
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=c4f2c1cd7dca1a65d877c4bfb5df15a829a1f521

+++ client/shared/Hooking.h
@@ -556,7 +556,7 @@ inline void call(AT address, T func)
 template<typename T>
 inline T get_call(T address)
 {
-	intptr_t target = *(uintptr_t*)(get_adjusted(address) + 1);
+	intptr_t target = *(int32_t*)(get_adjusted(address) + 1);
 	target += (get_adjusted(address) + 5);
 
 	return (T)target;
